requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
bandit==1.8.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
Django==5.2
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.21.2
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.8
Flask==3.1.0
Flask-PyMongo==2.3.0
fonttools==4.56.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyboard==0.13.5
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==2.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pbr==6.1.1
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyFirmata==1.1.0
pygame==2.6.0
Pygments==2.19.1
pymongo==4.8.0
pyparsing==3.2.1
pypiwin32==223
pyserial==3.5
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rich==13.9.4
ruff==0.11.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.31.0
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
stevedore==5.4.1
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
